PoC at implementing IntPtrUString via a common object
The added advantage is that every subslice of a given ustring will keep
a reference to the native pointer, so we have automatic memory
management for all the native values. Thus, we know that we can release
the handle safely even if it's shared between instances.
We also gain the advantage of the string wrappers being reclaimable,
only the shared object instance will remain in memory.